requirements.txt
matplotlib==3.7.1
mediapipe==0.9.2.1
moviepy==1.0.3
numpy==1.24.1
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
Pillow==9.5.0
scikit_learn==1.2.2
tensorboardX==2.6
tensorboardX==2.6
torch==2.0.0+cu118
torchvision==0.15.1+cu118
